pyproject.toml
[tool.poetry]
name = "kurs3"
version = "0.1.0"
description = ""
authors = ["Gleb Kozin <glebkozin7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-cov = "^5.0.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
